backend/requirements.txt
fastapi==0.96.0
uvicorn==0.22.0
telethon==1.29.2
python-dotenv==1.0.0
pydantic==1.10.8
starlette==0.27.0
websockets==11.0.3
cryptography==40.0.2
python-multipart==0.0.6
pandas==2.0.2
openpyxl==3.1.2
textblob==0.17.1
nltk==3.8.1
